Feature/single process compile (#778)
