+++ .drone.yml
@@ -5,8 +5,7 @@ pipeline:
     image: reimagined/resolve-docker
     pull: true
     commands:
-      - yarn install --ignore-scripts
-      - yarn prepare --parallel-limit=1
+      - yarn install
       - VERSION=$(/scripts/prepare_downstream)
       - /scripts/npm_patch_rc
       - npx oao all "/scripts/package_json patch-dependencies resolve $VERSION && /scripts/npm_publish $VERSION nightly"
