+++ .eslintrc.js
@@ -100,6 +100,7 @@ module.exports = {
           'deserialize',
           'deserialized',
           'dir',
+          'dirname',
           'dns',
           'dotenv',
           'downvote',
@@ -119,6 +120,7 @@ module.exports = {
           'fanout',
           'fas',
           'filename',
+          'filenames',
           'firebase',
           'firebaseio',
           'fff',
