--- babel.config.js
@@ -1,82 +1,0 @@
-module.exports = function(api) {
-  api.cache.using(
-    () => process.env.MODULE_TYPE + ':' + process.env.MODULE_TARGET
-  )
-
-  let useESModules, regenerator, helpers, modules, targets, loose
-
-  switch (process.env.MODULE_TYPE) {
-    case 'cjs': {
-      modules = 'commonjs'
-      useESModules = false
-      break
-    }
-    case 'es': {
-      modules = 'false'
-      useESModules = true
-      break
-    }
-    default:
-  }
-
-  switch (process.env.MODULE_TARGET) {
-    case 'server': {
-      loose = false
-      regenerator = false
-      helpers = false
-      targets = {
-        node: '8.10'
-      }
-      break
-    }
-    case 'client': {
-      loose = true
-      regenerator = true
-      helpers = true
-      break
-    }
-    default:
-  }
-
-  return {
-    presets: [
-      [
-        '@babel/preset-env',
-        {
-          loose,
-          targets,
-          modules
-        }
-      ],
-      '@babel/preset-react'
-    ],
-    plugins: [
-      ['@babel/plugin-proposal-decorators', { legacy: true }],
-      '@babel/plugin-proposal-class-properties',
-      '@babel/plugin-proposal-do-expressions',
-      '@babel/plugin-proposal-export-default-from',
-      '@babel/plugin-proposal-export-namespace-from',
-      '@babel/plugin-proposal-function-bind',
-      '@babel/plugin-proposal-function-sent',
-      '@babel/plugin-proposal-json-strings',
-      '@babel/plugin-proposal-logical-assignment-operators',
-      '@babel/plugin-proposal-nullish-coalescing-operator',
-      '@babel/plugin-proposal-numeric-separator',
-      '@babel/plugin-proposal-optional-chaining',
-      '@babel/plugin-proposal-object-rest-spread',
-      ['@babel/plugin-proposal-pipeline-operator', { proposal: 'minimal' }],
-      '@babel/plugin-proposal-throw-expressions',
-      '@babel/plugin-syntax-dynamic-import',
-      '@babel/plugin-syntax-import-meta',
-      [
-        '@babel/plugin-transform-runtime',
-        {
-          corejs: false,
-          helpers,
-          regenerator,
-          useESModules
-        }
-      ]
-    ]
-  }
-}
