+++ examples/hacker-news/package.json
@@ -43,7 +43,7 @@
     "resolve-storage-lite": "0.15.2",
     "sanitizer": "0.1.3",
     "seamless-immutable": "7.1.3",
-    "styled-components": "3.4.5",
+    "styled-components": "3.4.6",
     "uuid": "3.3.2"
   },
   "devDependencies": {
