+++ examples/with-styled-components/package.json
@@ -19,7 +19,7 @@
     "redux": "4.0.0",
     "resolve-redux": "0.15.2",
     "resolve-scripts": "0.15.2",
-    "styled-components": "3.4.5"
+    "styled-components": "3.4.6"
   },
   "devDependencies": {
     "@babel/core": "7.0.0",
