+++ jest.transform.js
@@ -1,7 +1,6 @@
-const babelrc = require('./babel.config')({
-  cache: {
-    using: () => {}
-  }
+const babelrc = require('./babel.compile').getConfig({
+  moduleType: 'cjs',
+  moduleTarget: 'server'
 })
 
 module.exports = require('babel-jest').createTransformer(babelrc)
