+++ package.json
@@ -44,7 +44,8 @@
     "prettier": "^1.14.2",
     "prettier-eslint-cli": "^4.7.1",
     "rimraf": "2.6.2",
-    "cross-env": "5.2.0"
+    "cross-env": "5.2.0",
+    "glob": "7.1.3"
   },
   "workspaces": {
     "packages": [
@@ -62,7 +63,7 @@
     "clean": "oao all \"npx rimraf dist lib es\" --parallel && npx rimraf yarn-error.log && npx oao clean",
     "lint": "eslint .",
     "lint-staged": "lint-staged",
-    "prepare": "oao all \"npx rimraf dist lib es\" --parallel && oao run-script compile --parallel",
+    "prepare": "node ./babel.compile.js --run",
     "prettier": "prettier --single-quote --no-semi --write \"**/**.js\"",
     "prettier-json": "prettier --write \"**/**.json\"",
     "publish": "oao publish -i \"examples/*\" --no-git-commit --no-changelog",
