+++ packages/adapters/bus-adapters/resolve-bus-memory/.npmignore
@@ -1,2 +1,3 @@
 test
 src
+compile.json
