+++ packages/adapters/bus-adapters/resolve-bus-rabbitmq/package.json
@@ -11,9 +11,6 @@
     "url": "https://github.com/reimagined/resolve/issues"
   },
   "scripts": {
-    "compile": "npm run compile:es && npm run compile:cjs",
-    "compile:es": "cross-env MODULE_TYPE=es MODULE_TARGET=server babel --config-file=../../../../babel.config.js --source-maps --out-dir ./es ./src",
-    "compile:cjs": "cross-env MODULE_TYPE=cjs MODULE_TARGET=server babel --config-file=../../../../babel.config.js --source-maps --out-dir ./lib ./src",
     "test": "jest --config=../../../../jest.config.js --verbose",
     "testw": "jest --config=../../../../jest.config.js --watchAll"
   },
@@ -33,5 +30,19 @@
   "devDependencies": {
     "jest": "23.6.0",
     "sinon": "6.2.0"
-  }
+  },
+  "babelCompile": [
+    {
+      "moduleType": "es",
+      "moduleTarget": "server",
+      "inputDir": "./src",
+      "outDir": "./es"
+    },
+    {
+      "moduleType": "cjs",
+      "moduleTarget": "server",
+      "inputDir": "./src",
+      "outDir": "./lib"
+    }
+  ]
 }
