+++ packages/adapters/subscribe-adapters/resolve-subscribe-socket.io/package.json
@@ -11,9 +11,6 @@
     "url": "https://github.com/reimagined/resolve/issues"
   },
   "scripts": {
-    "compile": "npm run build:client && npm run build:server",
-    "build:client": "cross-env MODULE_TYPE=cjs MODULE_TARGET=client babel --config-file=../../../../babel.config.js --source-maps --out-dir ./lib/client ./src/client",
-    "build:server": "cross-env MODULE_TYPE=cjs MODULE_TARGET=server babel --config-file=../../../../babel.config.js --source-maps --out-dir ./lib/server ./src/server",
     "test": "jest --config=../../../../jest.config.js --verbose",
     "testw": "jest --config=../../../../jest.config.js --watchAll"
   },
@@ -33,5 +30,19 @@
   "devDependencies": {
     "jest": "23.6.0"
   },
-  "homepage": "https://github.com/reimagined/resolve/tree/master/packages/adapters/subscribe-adapters/resolve-subscribe-socket.io#readme"
+  "homepage": "https://github.com/reimagined/resolve/tree/master/packages/adapters/subscribe-adapters/resolve-subscribe-socket.io#readme",
+  "babelCompile": [
+    {
+      "moduleType": "cjs",
+      "moduleTarget": "client",
+      "inputDir": "./src/client",
+      "outDir": "./lib/client"
+    },
+    {
+      "moduleType": "cjs",
+      "moduleTarget": "server",
+      "inputDir": "./src/server",
+      "outDir": "./lib/server"
+    }
+  ]
 }
