+++ packages/core/create-resolve-app/package.json
@@ -33,5 +33,13 @@
     "fs-extra": "^6.0.1",
     "https": "^1.0.0",
     "validate-npm-package-name": "^3.0.0"
-  }
+  },
+  "babelCompile": [
+    {
+      "moduleType": "cjs",
+      "moduleTarget": "server",
+      "inputDir": "./src",
+      "outDir": "./lib"
+    }
+  ]
 }
