+++ packages/core/resolve-scripts/package.json
@@ -17,9 +17,6 @@
     "ddd"
   ],
   "scripts": {
-    "compile": "npm run compile:es && npm run compile:cjs",
-    "compile:es": "cross-env MODULE_TYPE=es MODULE_TARGET=server babel --config-file=../../../babel.config.js --source-maps --out-dir ./es ./src",
-    "compile:cjs": "cross-env MODULE_TYPE=cjs MODULE_TARGET=server babel --config-file=../../../babel.config.js --source-maps --out-dir ./lib ./src",
     "test": "jest --config=../../../jest.config.js --verbose",
     "testw": "jest --config=../../../jest.config.js --watchAll"
   },
@@ -94,5 +91,19 @@
   },
   "optionalDependencies": {
     "styled-components": ">=3.4.5"
-  }
+  },
+  "babelCompile": [
+    {
+      "moduleType": "es",
+      "moduleTarget": "server",
+      "inputDir": "./src",
+      "outDir": "./es"
+    },
+    {
+      "moduleType": "cjs",
+      "moduleTarget": "server",
+      "inputDir": "./src",
+      "outDir": "./lib"
+    }
+  ]
 }
